Refactor first_record function to move it from lookup to search module for better organization
